Make MinHash hash values consistent across different Python versions (#24)
* Test minhash against fixed result vector

I am not sure if your MinHash implementation is supposed to give the  same results on different versions of Python. It currently does not do so. I added the test vectors for the results I get on python 2.7. Unfortunately Python 3.5 gives different results.

* Replace inconsistent random with numpy random

After some research it turned out that the standard lib pseudo random generation is inconsistent across python versions. Switching to the numpy provided random generator fixes the problem

* Update test vector to new numpy random generator

* Use deterministic pseudo-random generator for benchmark

To be able to compare benchmarks between py27/py35 we need to test with the same data.